requirements.txt
Django~=5.1
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
drf-yasg~=1.21.7
psycopg2~=2.9.9
psycopg2-binary~=2.9.9
setuptools~=74.1.2
pillow~=10.4.0